Create README.md
Added the travis.yaml

Update README.md

added the jetty xml

added spock test

+++ Jersey2-Spring4-Servlet-3-Spock/src/main/java/org/harry/rs/employeesample/dao/EmployeeDAO.java
@@ -1,7 +1,6 @@
 package org.harry.rs.employeesample.dao;
 
 import org.harry.rs.employeesample.entity.EmployeeEntity;
-import org.harry.rs.employeesample.model.Employee;
 
 import java.util.List;
 
@@ -15,5 +14,5 @@ public interface EmployeeDAO {
     public List<EmployeeEntity>  saveEmployees(List<EmployeeEntity> ems);
     public EmployeeEntity saveEmployee(EmployeeEntity em);
 
-    void delete(Integer id);
+    void delete(Long id);
 }
